consolidate api related code into the api lib module
